requirements.txt
torch>=2.0.0
numpy>=1.24.0,<2.0.0
fastf1>=3.4.0
langgraph>=0.2.0
deap>=1.4.0
groq>=0.4.0
httpx>=0.25.0
influxdb-client>=1.36.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
